refactor: get doodle illustrations from backend using media collection mapping
Previously doodle illustrations had to be in the frontend in the path src/assets/images/verk/, now they must be in the backend. Use collections.mediaCollectionMappings in config.ts to map the id of the collection with doodles to the id of the media collection that holds the images in the backend.