feature: add config option to replace image assets paths in collection texts
`config.collections.replaceImageAssetsPaths`.
The new config option defaults to `true`, which corresponds to the old behaviour. When this option is `true`, `src` attribute values starting with `images/` in the HTML of collection texts are replaced with `assets/images/`, which is the correct path to the image assets folder. Setting the new config option to `false` improves app performance.