fix: broken SSR for non-default locales
Before this commit pages in non-default locales that used the `UrlService` were not rendered on the server. This was a side effect of the `@jsonurl/jsonurl` library used by the service to convert primitive data types and complex data structures like arrays, arrays of objects and objects with nested arrays and objects into URL-safe strings and vice versa. This fix removes the dependency on the external library and replaces its functionality with a custom conversion implementation. The custom implementation is not a full implementation of the JSON→URL specification, but produces the same results for the parts that are relevant for this project. The code has been generated by ChatGPT-4 according to the JSON→URL specification, meticulously guided by a human mind.