fix: preserve selected facsimile collection on text change
When navigating from one collection text to another and you have a facsimile view open, a facsimile from the selected facsimile collection is opened if applicable to the new text. The preservation is performed based on the priority/sortOrder of the facsimiles linked to the text. So if a text has two linked facsimiles, with priorities 1 and 2, each opened in a view, and the user navigates to the next text, the facsimile views are loaded with facsimiles with priorities 1 and 2, if the new text has such facsimiles.